Split operator across different files.
This PR split the operator.go file to different files based on scope.
