+++ pkg/operator/kapacitor.go
@@ -1,0 +1,102 @@
+package operator
+
+import (
+	"fmt"
+	"log"
+
+	"github.com/gianarb/influxdb-operator/pkg/client/tick/v1alpha1"
+	"k8s.io/api/apps/v1beta1"
+	"k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/client-go/tools/cache"
+)
+
+func registerKapacitorInformer(operator *Operator) {
+	operator.kapacitorInformer = cache.NewSharedIndexInformer(
+		&cache.ListWatch{
+			ListFunc:  operator.tickCs.Kapacitors(metav1.NamespaceAll).List,
+			WatchFunc: operator.tickCs.Kapacitors(metav1.NamespaceAll).Watch,
+		},
+		&v1alpha1.Kapacitor{}, 0, cache.Indexers{},
+	)
+
+	operator.kapacitorInformer.AddEventHandler(cache.ResourceEventHandlerFuncs{
+		AddFunc:    operator.handleAddKapacitor,
+		UpdateFunc: nil,
+		DeleteFunc: operator.handleDeleteKapacitor,
+	})
+}
+
+func (o *Operator) handleDeleteKapacitor(obj interface{}) {
+	oret, ok := o.getObject(obj)
+
+	if !ok {
+		//TODO: error? panic? how?
+		return
+	}
+
+	deploymentName := fmt.Sprintf("%s-%s", v1alpha1.KapacitorPlural, oret.GetName())
+	policy := metav1.DeletePropagationForeground
+	err := o.kubeClient.ExtensionsV1beta1().Deployments(oret.GetNamespace()).Delete(deploymentName, &metav1.DeleteOptions{
+		PropagationPolicy: &policy,
+	})
+	if err != nil {
+		log.Printf("Error deleting deployment %s. %s", deploymentName, err)
+	}
+}
+
+func (o *Operator) handleAddKapacitor(obj interface{}) {
+	oret, ok := o.getObject(obj)
+
+	if !ok {
+		//TODO: error? panic? how?
+		return
+	}
+
+	kapacitorSpec := obj.(*v1alpha1.Kapacitor)
+
+	labels := o.config.Labels
+	labels["name"] = fmt.Sprintf("%s-%s", v1alpha1.KapacitorPlural, oret.GetName())
+	labels["resource"] = v1alpha1.KapacitorPlural
+
+	replicas := int32(1)
+	deployment := &v1beta1.Deployment{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      labels["name"],
+			Labels:    labels,
+			Namespace: oret.GetNamespace(),
+		},
+		Spec: v1beta1.DeploymentSpec{
+			Replicas: &replicas,
+			Template: v1.PodTemplateSpec{
+				ObjectMeta: metav1.ObjectMeta{
+					Labels:    labels,
+					Namespace: oret.GetNamespace(),
+				},
+				Spec: v1.PodSpec{
+					Containers: []v1.Container{
+						v1.Container{
+							Name:            oret.GetName(),
+							Image:           kapacitorSpec.Spec.BaseImage,
+							ImagePullPolicy: "Always",
+							Env:             []v1.EnvVar{},
+							Ports: []v1.ContainerPort{
+								v1.ContainerPort{
+									Name:          "http",
+									ContainerPort: 9092,
+									Protocol:      v1.ProtocolTCP,
+								},
+							},
+							//VolumeMounts: []v1.VolumeMount{},
+							//Resources:    v1.ResourceRequirements{},
+						},
+					},
+					//Volumes: []v1.Volume{},
+				},
+			},
+		},
+	}
+
+	_, err := o.kubeClient.AppsV1beta1().Deployments(oret.GetNamespace()).Create(deployment)
+	fmt.Print(err) //TODO: handle in a different way?
+}
